- new folder structure - log4j -> slf4j + logback - spring boot dependencies - clean up duplicates in the dependencies
