--- dao-api/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <artifactId>hr</artifactId>
-        <groupId>com.epam.brest</groupId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>dao-api</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.epam.brest</groupId>
-            <artifactId>model</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
