+++ dao/api/pom.xml
@@ -3,15 +3,22 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <artifactId>dao-api</artifactId>
     <packaging>jar</packaging>
-    <artifactId>model</artifactId>
-    <name>model</name>
-    <!-- TODO: do we need split DTO and Entities? -->
-    <description>DTOs for web, entities for the database</description>
+    <name>dao-api</name>
+    <description>DAO interfaces</description>
 
     <parent>
+        <artifactId>dao</artifactId>
         <groupId>com.epam.brest</groupId>
-        <artifactId>hr</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.epam.brest</groupId>
+            <artifactId>model</artifactId>
+        </dependency>
+    </dependencies>
+
 </project>
