--- dao/src/main/resources/log4j2.properties
@@ -1,12 +1,0 @@
-status=warn
-name=PropertiesConfig
-appenders=console
-appender.console.type=Console
-appender.console.name=STDOUT
-appender.console.layout.type=PatternLayout
-appender.console.layout.pattern=%d{HH:mm:ss} %-5p %-20.20C{1} %m%n
-log.test_app.name=com.epam.brest.dao
-log.test_app.level=debug
-log.test_app.appenderRef.stdout.ref=STDOUT
-rootLogger.level=debug
-rootLogger.appenderRef.stdout.ref=STDOUT
