--- dao/src/test/resources/ log4j2-test.properties
@@ -1,9 +1,0 @@
-status=debug
-name=PropertiesConfig
-appenders=console
-appender.console.type=Console
-appender.console.name=STDOUT
-appender.console.layout.type=PatternLayout
-appender.console.layout.pattern=%d{HH:mm:ss} %-5p %-20.20C{1} %m%n
-rootLogger.level=debug
-rootLogger.appenderRef.stdout.ref=STDOUT
