+++ rest-app/pom.xml
@@ -12,31 +12,11 @@
     <artifactId>rest-app</artifactId>
 
     <dependencies>
-
-        <!--        <dependency>-->
-        <!--            <groupId>com.epam.brest</groupId>-->
-        <!--            <artifactId>model</artifactId>-->
-        <!--        </dependency>-->
-        <!--        <dependency>-->
-        <!--            <groupId>com.epam.brest</groupId>-->
-        <!--            <artifactId>dao-api</artifactId>-->
-        <!--        </dependency>-->
-        <!--        <dependency>-->
-        <!--            <groupId>com.epam.brest</groupId>-->
-        <!--            <artifactId>dao</artifactId>-->
-        <!--        </dependency>-->
-        <!--        <dependency>-->
-        <!--            <groupId>com.epam.brest</groupId>-->
-        <!--            <artifactId>service-api</artifactId>-->
-        <!--        </dependency>-->
         <dependency>
             <groupId>com.epam.brest</groupId>
-            <artifactId>service</artifactId>
+            <artifactId>service-impl</artifactId>
+            <version>${project.version}</version>
         </dependency>
-        <!--        <dependency>-->
-        <!--            <groupId>com.epam.brest</groupId>-->
-        <!--            <artifactId>test-db</artifactId>-->
-        <!--        </dependency>-->
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -48,9 +28,11 @@
             <artifactId>spring-boot-starter-actuator</artifactId>
         </dependency>
 
+        <!-- test -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
+            <scope>test</scope>
         </dependency>
 
         <dependency>
