+++ service-rest/src/main/java/com/epam/brest/service/rest/DepartmentDtoServiceRest.java
@@ -2,8 +2,8 @@ package com.epam.brest.service.rest;
 
 import com.epam.brest.model.dto.DepartmentDto;
 import com.epam.brest.service.DepartmentDtoService;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
@@ -13,7 +13,7 @@ import java.util.List;
 @Service
 public class DepartmentDtoServiceRest implements DepartmentDtoService {
 
-    private final Logger logger = LogManager.getLogger(DepartmentDtoServiceRest.class);
+    private final Logger logger = LoggerFactory.getLogger(DepartmentDtoServiceRest.class);
 
     private String url;
