+++ service/api/pom.xml
@@ -3,15 +3,21 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <artifactId>service-api</artifactId>
     <packaging>jar</packaging>
-    <artifactId>model</artifactId>
-    <name>model</name>
-    <!-- TODO: do we need split DTO and Entities? -->
-    <description>DTOs for web, entities for the database</description>
+    <name>service-api</name>
+    <description>Service interfaces only</description>
 
     <parent>
+        <artifactId>service</artifactId>
         <groupId>com.epam.brest</groupId>
-        <artifactId>hr</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.epam.brest</groupId>
+            <artifactId>dao-impl</artifactId>
+        </dependency>
+    </dependencies>
 </project>
