+++ service/impl/src/main/java/com/epam/brest/service/exceptions/DepartmentNotFoundException.java
