+++ service/impl/src/main/java/com/epam/brest/service/exceptions/package-info.java
